+++ src/components/Table/Table.tsx
@@ -14,10 +14,11 @@ import {
   getPaginationRowModel,
   getSortedRowModel,
   SortingState,
+  Table as TableInstance,
   useReactTable,
 } from '@tanstack/react-table';
 import { useVirtualizer } from '@tanstack/react-virtual';
-import React, { CSSProperties, RefObject, useCallback, useMemo, useState } from 'react';
+import React, { CSSProperties, MutableRefObject, RefObject, useCallback, useEffect, useMemo, useState } from 'react';
 
 import { TEST_IDS } from '@/constants';
 import { useEditableData, useSyncedColumnFilters } from '@/hooks';
@@ -109,6 +110,11 @@ interface Props<TData> {
    * @type {PaginationOptions}
    */
   pagination: PaginationOptions;
+
+  /**
+   * Table Instance
+   */
+  tableInstance: MutableRefObject<TableInstance<TData>>;
 }
 
 /**
@@ -167,6 +173,7 @@ export const Table = <TData,>({
   paginationRef,
   width,
   pagination,
+  tableInstance,
 }: Props<TData>) => {
   /**
    * Styles and Theme
@@ -301,6 +308,13 @@ export const Table = <TData,>({
    * Editable Data
    */
   const editableData = useEditableData({ table, onUpdateRow });
+
+  /**
+   * Set table instance
+   */
+  useEffect(() => {
+    tableInstance.current = table;
+  }, [table, tableInstance]);
 
   return (
     <>